action plans have a copy of all the plan's parameter
a copy of the parameter is created for each plan instead of only the
action's parameters because the predicates in the action may use
parameters that are not one of the action's parameter.

a symbol is assumed to be a consensus variable so that computation
doesn't need to be done to determine this.